[feat] 抽出模块、优化、添加@EntityField注解、主键由Long改成Integer

+++ wonder-admin/src/main/java/com/gitee/whzzone/admin/business/service/OrderService.java
@@ -3,7 +3,7 @@ package com.gitee.whzzone.admin.business.service;
 import com.gitee.whzzone.admin.business.entity.Order;
 import com.gitee.whzzone.admin.business.pojo.dto.OrderDto;
 import com.gitee.whzzone.admin.business.pojo.query.OrderQuery;
-import com.gitee.whzzone.common.base.service.EntityService;
+import com.gitee.whzzone.web.service.EntityService;
 
 import java.math.BigDecimal;
 import java.util.List;
@@ -15,6 +15,6 @@ public interface OrderService extends EntityService<Order, OrderDto, OrderQuery>
 
     List<OrderDto> list(OrderQuery query);
 
-    List<Long> limitAmountBetween(BigDecimal begin, BigDecimal end);
+    List<Integer> limitAmountBetween(BigDecimal begin, BigDecimal end);
 
 }
