+++ wonder-admin/src/main/java/com/gitee/whzzone/admin/system/controller/DeptController.java
@@ -1,11 +1,11 @@
 package com.gitee.whzzone.admin.system.controller;
 
-import com.gitee.whzzone.common.base.controller.EntityController;
-import com.gitee.whzzone.admin.system.pojo.dto.DeptDto;
 import com.gitee.whzzone.admin.system.entity.Dept;
+import com.gitee.whzzone.admin.system.pojo.dto.DeptDto;
 import com.gitee.whzzone.admin.system.pojo.query.DeptQuery;
 import com.gitee.whzzone.admin.system.service.DeptService;
-import com.gitee.whzzone.web.Result;
+import com.gitee.whzzone.web.controller.EntityController;
+import com.gitee.whzzone.web.pojo.other.Result;
 import io.swagger.annotations.Api;
 import io.swagger.annotations.ApiOperation;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -39,7 +39,7 @@ public class DeptController extends EntityController<Dept, DeptService, DeptDto,
 
     @ApiOperation("改变启用状态")
     @GetMapping("/enabledSwitch/{id}")
-    public Result enabledSwitch(@PathVariable Long id) {
+    public Result enabledSwitch(@PathVariable Integer id) {
         deptService.enabledSwitch(id);
         return Result.ok();
     }
