+++ wonder-admin/src/main/java/com/gitee/whzzone/admin/system/controller/RuleController.java
@@ -1,11 +1,11 @@
 package com.gitee.whzzone.admin.system.controller;
 
-import com.gitee.whzzone.common.base.controller.EntityController;
+import com.gitee.whzzone.web.controller.EntityController;
 import com.gitee.whzzone.admin.system.pojo.dto.RuleDto;
 import com.gitee.whzzone.admin.system.entity.Rule;
 import com.gitee.whzzone.admin.system.pojo.query.RuleQuery;
 import com.gitee.whzzone.admin.system.service.RuleService;
-import com.gitee.whzzone.web.Result;
+import com.gitee.whzzone.web.pojo.other.Result;
 import io.swagger.annotations.Api;
 import io.swagger.annotations.ApiOperation;
 import lombok.extern.slf4j.Slf4j;
@@ -31,13 +31,13 @@ public class RuleController extends EntityController<Rule, RuleService, RuleDto,
 
     @ApiOperation("根据markId获取ruleList")
     @GetMapping("getByMarkId/{markId}")
-    public Result<List<RuleDto>> getByMarkId(@PathVariable Long markId){
+    public Result<List<RuleDto>> getByMarkId(@PathVariable Integer markId){
         return Result.ok(ruleService.getByMarkId(markId));
     }
 
     @ApiOperation("根据roleId，markId获取一个rule")
     @GetMapping("getByRoleIdAndMarkId")
-    public Result<RuleDto> getByRoleIdAndMarkId(Long roleId, Long markId){
+    public Result<RuleDto> getByRoleIdAndMarkId(Integer roleId, Integer markId){
         return Result.ok(ruleService.getByRoleIdAndMarkId(roleId, markId));
     }
