+++ wonder-admin/src/main/java/com/gitee/whzzone/admin/system/entity/RequestLog.java
@@ -2,7 +2,7 @@ package com.gitee.whzzone.admin.system.entity;
 
 import com.baomidou.mybatisplus.annotation.TableField;
 import com.baomidou.mybatisplus.annotation.TableName;
-import com.gitee.whzzone.common.base.pojo.entity.BaseEntity;
+import com.gitee.whzzone.web.entity.BaseEntity;
 import io.swagger.annotations.ApiModel;
 import io.swagger.annotations.ApiModelProperty;
 import lombok.Getter;
@@ -26,7 +26,7 @@ public class RequestLog extends BaseEntity {
 
     @ApiModelProperty("请求人")
     @TableField("user_id")
-    private Long userId;
+    private Integer userId;
 
     @ApiModelProperty("请求url")
     @TableField("url")
