+++ wonder-admin/src/main/java/com/gitee/whzzone/admin/system/service/MenuService.java
@@ -1,6 +1,6 @@
 package com.gitee.whzzone.admin.system.service;
 
-import com.gitee.whzzone.common.base.service.EntityService;
+import com.gitee.whzzone.web.service.EntityService;
 import com.gitee.whzzone.admin.system.pojo.dto.MenuTreeDto;
 import com.gitee.whzzone.admin.system.pojo.dto.MenuDto;
 import com.gitee.whzzone.admin.system.entity.Menu;
@@ -13,7 +13,7 @@ import java.util.List;
  * @date : 2023/5/22 20:17
  */
 public interface MenuService  extends EntityService<Menu, MenuDto, MenuQuery> {
-    List<String> findPermitByUserId(Long userId);
+    List<String> findPermitByUserId(Integer userId);
 
     List<MenuTreeDto> treeList(MenuQuery query);
 
@@ -21,11 +21,11 @@ public interface MenuService  extends EntityService<Menu, MenuDto, MenuQuery> {
 
     List<Menu> getEnabledList();
 
-    List<MenuDto> findByUserId(Long userId);
+    List<MenuDto> findByUserId(Integer userId);
 
-    List<Long> getIdListByRoleId(Long id);
+    List<Integer> getIdListByRoleId(Integer id);
 
-    boolean existSamePermission(Long id, String Permission);
+    boolean existSamePermission(Integer id, String Permission);
 
-    boolean existSameRouteName(Long id, String routeName);
+    boolean existSameRouteName(Integer id, String routeName);
 }
