+++ wonder-admin/src/main/java/com/gitee/whzzone/admin/system/service/RoleMarkService.java
@@ -1,6 +1,6 @@
 package com.gitee.whzzone.admin.system.service;
 
-import com.gitee.whzzone.common.base.service.EntityService;
+import com.gitee.whzzone.web.service.EntityService;
 import com.gitee.whzzone.admin.system.pojo.dto.RoleMarkDto;
 import com.gitee.whzzone.admin.system.entity.RoleMark;
 import com.gitee.whzzone.admin.system.pojo.query.RoleMarkQuery;
@@ -12,8 +12,8 @@ import java.util.List;
  */
 public interface RoleMarkService extends EntityService<RoleMark, RoleMarkDto, RoleMarkQuery> {
 
-    List<RoleMark> getByRoleId(Long roleId);
+    List<RoleMark> getByRoleId(Integer roleId);
 
-    List<RoleMark> getByRoleIdAndMarkId(Long roleId, Long ruleId);
+    List<RoleMark> getByRoleIdAndMarkId(Integer roleId, Integer ruleId);
 
 }
