+++ wonder-admin/src/main/java/com/gitee/whzzone/admin/system/service/UserRoleService.java
@@ -1,6 +1,6 @@
 package com.gitee.whzzone.admin.system.service;
 
-import com.gitee.whzzone.common.base.service.EntityService;
+import com.gitee.whzzone.web.service.EntityService;
 import com.gitee.whzzone.admin.system.pojo.dto.UserRoleDto;
 import com.gitee.whzzone.admin.system.entity.UserRole;
 import com.gitee.whzzone.admin.system.pojo.query.UserRoleQuery;
@@ -13,6 +13,6 @@ import java.util.List;
 
 public interface UserRoleService extends EntityService<UserRole, UserRoleDto, UserRoleQuery> {
 
-    List<UserRole> getByUserId(Long userId);
+    List<UserRole> getByUserId(Integer userId);
 
 }
